Speed change and shop pause UI start

+++ Assets/Scripts/EnemyStats.cs
@@ -8,7 +8,7 @@ public class EnemyStats : Stats
 	public float speed = 2f;
 	public float attackRange = 3f;
 	public float damage = 3f;
-	public float value = 3f;
+	public uint value = 3;
 	public ProjectileData projectileData;
 	public float attackSpeed;
 }
